ci: use NPM packages when generating
